pyproject.toml
[project]
name = "cli-agent"
version = "0.0.1"
description = "A helpful RAG agent living in the CLI capable of web scraping and automatic knowledge ingestion."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cutie>=0.3.2",
    "deepagents==0.0.5",
    "langchain-mcp-adapters==0.1.10",
    "langchain-openai==0.3.33",
    "loguru>=0.7.3",
    "pixeltable>=0.4.14",
    "rich>=14.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cli_agent"]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "tavily-python>=0.7.12",
]
